requirements.txt
albumentations==0.5.2
alembic==1.5.4
anyio==2.1.0
appdirs==1.4.3
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==3.3.0
CacheControl==0.12.6
certifi==2019.11.28
certipy==0.1.3
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
efficientnet-pytorch==0.6.3
entrypoints==0.3
html5lib==1.0.1
idna==2.8
imageio==2.9.0
imgaug==0.4.0
ipaddr==2.2.0
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyter-server==1.3.0
jupyter-telemetry==0.1.0
jupyterlab==3.0.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.2.0
kiwisolver==1.3.1
lockfile==0.12.2
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
msgpack==0.6.2
munch==2.5.0
natsort==7.1.1
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==20.3
pamela==1.0.0
pandocfilters==1.4.3
parso==0.8.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pretrainedmodels==0.7.4
progress==1.5
prometheus-client==0.9.0
prompt-toolkit==3.0.14
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.7.4
pyOpenSSL==20.0.1
pyparsing==2.4.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-editor==1.0.4
python-json-logger==2.0.1
pytoml==0.1.21
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.2
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
segmentation-models-pytorch==0.1.3
Send2Trash==1.5.0
Shapely==1.7.1
six==1.14.0
sklearn==0.0
sniffio==1.2.0
SQLAlchemy==1.3.23
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.2.1
timm==0.3.2
torchvision
tornado==6.1
tqdm==4.56.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
